Change flowtype file hierarchy for tests
